phi_simplification_pass: Run local CSE after deduping SCCs
This can expose more opt possiblities
